requirements.txt
yfinance
boto3
python-dotenv